refactor: #55 modify repository file tree
