+++ src/user/user.module.ts
@@ -1,10 +1,10 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { UserRepository } from './user.repository';
 import { UserService } from './user.service';
 import { UserController } from './user.controller';
 import { UserProfileService } from './user-profile.service';
-import { UserProfileRepository } from './user-profile.repository';
+import { UserProfileRepository } from './repository/user-profile.repository';
+import { UserRepository } from './repository/user.repository';
 
 @Module({
   imports: [
